Move normalize-fields.ts to utils/normalize-fields.ts
